make flags component reusable
